requirements.txt
requests
python-dotenv
pytz
cloudscraper
# Pin discord.py to a version compatible with Python 3.13
discord.py==2.4.0
aiosqlite>=0.19.0
aiohttp>=3.9.0